# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "kinoage",
        "description": "Documentation for kinoage admin dashboard and main site",
        "contact": {
            "email": "ayubxonnematov@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/movie": {
            "get": {
                "tags": [
                    "Movies"
                ],
                "summary": "Get list of Movie",
                "description": "Returns list of Movie",
                "operationId": "indexMovie",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Movie"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Movies"
                ],
                "summary": "Insert a new Movie",
                "description": "Insert a new Movie",
                "operationId": "storeMovie",
                "requestBody": {
                    "description": "Movie to create",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "data": {
                                        "$ref": "#/components/schemas/Movie"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Movie created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Movie"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation exception"
                    }
                }
            }
        },
        "/api/movie/{movie_id}": {
            "get": {
                "tags": [
                    "Movies"
                ],
                "summary": "Show a Movie from his Id",
                "description": "Show a Movie from his Id",
                "operationId": "showMovie",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Movie--id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Movie"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Movie not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Movies"
                ],
                "summary": "Delete a Movie",
                "description": "Delete a Movie",
                "operationId": "destroyMovie",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Movie--id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Movie"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Movie not found"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Movies"
                ],
                "summary": "Update an existing Movie",
                "description": "Update an existing Movie",
                "operationId": "updateMovie",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Movie--id"
                    }
                ],
                "requestBody": {
                    "description": "Movie to update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "data": {
                                        "$ref": "#/components/schemas/Movie"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Movie"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Movie not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Movie": {
                "title": "Movie",
                "description": "Movie model",
                "required": [],
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id of Movie",
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "name": {
                        "title": "name",
                        "description": "name of Movie",
                        "type": "string",
                        "example": "Macbook Pro"
                    },
                    "sku": {
                        "title": "sku",
                        "description": "sku of Movie",
                        "type": "string",
                        "example": "MCBPRO2022"
                    },
                    "price": {
                        "title": "price",
                        "description": "price of Movie",
                        "type": "integer",
                        "example": "99"
                    },
                    "created_at": {
                        "title": "created_at",
                        "type": "dateTime",
                        "readOnly": "true",
                        "example": "2022-07-04T02:41:42.336Z"
                    },
                    "updated_at": {
                        "title": "updated_at",
                        "type": "dateTime",
                        "readOnly": "true",
                        "example": "2022-07-04T02:41:42.336Z"
                    }
                },
                "type": "object"
            },
...
        },
        "parameters": {
...
            "Movie--id": {
                "name": "Movie_id",
                "in": "path",
                "description": "Id of Movie",
                "required": true,
                "schema": {
                    "type": "integer",
                    "example": "1"
                }
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}